Realized I wasn't setting a background for all windows, got that fixed. Also realized that the values in /proc/[pid]/status aren't always the same between kernel versions, so I added a null check when displaying a processes status. Also adjusted the capacity of a few arenas to save some space. If the process status window information is to be trusted it didn't make much of a dent. If the program starts crashing after any period of time I probably need to revert that section of code to the previous commit and start over.

+++ src/monitor/prc_monitor.c
@@ -11,7 +11,7 @@
 #define MAX_PROC_REGIONS_ALLOCD 3
 
 
-static int _copy_if_tracked_stat(char *buf);
+static u8 _copy_if_tracked_stat(char *buf);
 
 static void _fetch_proc_pid_stat(
     Arena *prcArena,
@@ -166,7 +166,7 @@ void get_prc_info_by_pid(volatile ProcessInfoSharedData *prcInfoSd)
 	if (sscanf(statusBuffer, "Name:\t%s", prcInfoSd->info->procName) > 0) continue;
 	else if (sscanf(statusBuffer, "Pid:\t%d", &prcInfoSd->info->pid) > 0) continue;
 
-	int isTracked = _copy_if_tracked_stat(statusBuffer);
+	u8 isTracked = _copy_if_tracked_stat(statusBuffer);
 
 	if (isTracked) 
 	{
@@ -177,7 +177,7 @@ void get_prc_info_by_pid(volatile ProcessInfoSharedData *prcInfoSd)
     fclose(statusFile);
 }
 
-static int _copy_if_tracked_stat(char *buf)
+static u8 _copy_if_tracked_stat(char *buf)
 {
     char **res = bsearch(buf, trackedStats, 19, sizeof(char *), prc_tracked_stat_cmp);
